Variable 'keystroke' is used uninitialized whenever function 'nwipe_gui_prng' is called.
The consequence of this uninitialised variable will be that intermittently pressing pressing 'P' won't do anything, that will occur if the uninitialised 'keystroke' happens to be equal to the defined value of ERR used in the while statement. This may or maynot be an extremely rare bug depending upon the unitialised contents of 'keystroke'. The fix is instead of using a while() loop, a do{}while() loop is used. Located in the function nwipe_gui_prng( void ) in gui.c