Fix individual device throughput.
On completion of a device wipe, the throughput was continuing
to be calculated for a completed wipe which resulted in the
throughput for that particular device to gradually drop over time
to zero if other drives were still

This has now been corrected. Throughput calculation for a given
device is stopped when the particular device has completed its wipe.