Check for dmidecode in /sbin
In Debian Sid dmidecode is not found as /sbin is not
searched when nwipe only specifies 'dmidecode', so
if it fails then it will next try /sbin/dmidecode
which is where it's normally located if it's installed.

In ubuntu etc, the sudo environment is setup so /sbin
is automatically searched.