Warn user about no drives selected
If a user attempts to start a wipe but hasn't yet
selected any drives for wiping, a warning message
will now appear for a few seconds suggesting
they select a drive.

Prior to this patch, if you tried to start a wipe
without selecting a drive, nwipe would immediately
display a message saying the wipe was complete
and press enter to exit. You had no choice but to
exit and restart nwipe.