Fix_error_reporting_on_ctrl_C before wipe starts
If you do a ctrl c to exit nwipe before you have
even started a wipe, it will be reported in the logs
that there are fatal or non fatal errors. This is
incorrect as no wipe has even been started.

This error is only relevant to when using ctrl c
before a wipe has started and not after a wipe has
started.

Basically the variables that hold the errors are
in an indeterminate state before wiping begins and
in the case of a user abort we mistakenly checked
those values and reported. They should only be used
after a thread has completed or failed when their
values are then valid.

Now if you exit nwipe before any wiping started
there are no fatal/non fatal errors reported.