Moved the temperature readout
In both selection and wipe status windows I
moved the temperature readout to the right
of the disk size column.

Also in the wipe status window, by moving to
the right of the disk size column the
temperature is no longer on the same line as
the disk pass progress. This reduces the line
length as I found that in tests if you had
many verification errors the temperature
being on the end of the line would disappear
of the right of the screen.