Fix obscure segfault when --logfile option used.
If you use the --logfile option but specify a filename
that will be created in a directory that is not writable,
for instance, a system directory such as /proc/sys/ or
your current directory happens to be a system directory
that you are running nwipe from while not specifying a
writable path for the log file then nwipe would exit with
a segfault.

This is now fixed and if the logfile cannot be created
or opened then an appropriate message is displayed and
nwipe is aborted.