PDFGen11 - further work on PDF certificate
1. Fix a bug in the static and random pass code
where I unnecessarily moved the decrement of z (the
count of bytes remaining) to after the periodic sync
which meant every time it sync'ed z wasn't decremented due
to r being used as a general results variable for the
fdatasync function.

I prefer to just have a variable be responsible for just
one thing hence my assumption r wasn't used for two different
things.

Changed the position of the decrement of z, pass_done and
round_done to their original position prior to the periodic
sync.